widget: add gtk_widget_set_action_parent()
This adds a new function to the 4.8 ABI that allows setting an action
parent for a widget. The action parent affects the action muxing so that
instead of using the widgets direct ancestor, an alternate widget's
action muxer may be used.

You may not set an action parent for a widget that is a direct descendant
of your widget as that would cause cycles in action resolution.

You might find this API useful for situations where you want menus in
headerbars to route through action muxers for the current document as
well as toolbars or sidebars.

Fixes #4860