CI: Use mutter for the x11-full testsuite run
This spawns a mutter WM for the xvfb instance and wraps it in its own
dbus session.

Also use a larger screen size (1080p) - mostly so we can check the
testuite runs with differently sized screens, but alspo because that's
today's default screen size.

This exposes the issue from #4136 in the testsuite.