gtkimcontextime.c: Port to Pango2 APIs
Update the code to use Pango2 APIs, by removing includes to pangowin32, which
is gone, and replacing relevant calls to Pango2 equivilants.

Replace the call to pango_win32_font_logfontw() by retrieving instead the
underlying hb_font_t, and calling hb_uniscribe_font_get_logfontw(), but preedit
fonts are currently not working until a resolution to HarfBuzz issue #3683 can
be found.  Since we are still using Windows IME APIs for input, we need the
HarfBuzz GDI/Uniscribe items to be included in the HarfBuzz builds.