build: fix check for debug vs. release builds
Confusingly, Meson's debug option indicates whether debuginfo is
enabled, not whether we have a debug build. We should consider the
optimization level as well when deciding whether to enable debug extras.

I've outlined an argument for why we should make these changes here:

https://blogs.gnome.org/mcatanzaro/2022/07/15/best-practices-for-build-options/

The main take-away is this rule:

“You have a non-production (debug) build if debug is true and if
optimization is 0 or g; otherwise, you have a production build” (except
on Windows, which is not accounted for here).

This commit additionally removes guidance related to manually
controlling preprocessor definitions in plain builds, which is
controversial and violates Rule 6 from my blog post.