win32: Add NMake Makefile to bootstrap projects
This adds a NMake Makefile to generate the needed items that are normally
generated during `./configure`, as well as to generate the full Visual Studio
projects and property sheets, so that one can use the Visual Studio projects
to build GTK directly from a GIT checkout.  A 'clean' target is also provided
to remove these generated artifacts.  This also generates the headers that are
needed to build broadwayd, since they are currently dist'ed with the tarballs
but are not in the GIT repo.

This will not be dist'ed, as this is primarily meant for GIT checkouts only.