Move focus to parent when hiding a focused child
because otherwise the focus will jump forward
to next widget, while expected behaviour is
for focus to come back to parent widget when
such parent is focusable.

As we now do this toolkit wide in
_gtk_window_unset_focus_and_default(),
there's no need for individual widgets to do
that explicitly, so we remove such explicit
calls by the GtkText and GtkTextView widgets.

Fixes issue #4903