Prioritize scrolling a GtkScrolledWindow over its child widgets
Scrolling a GtkScrolledWindow should have precedence
over child widgets that also react to scrolling
like GtkComboBox, GtkScale and GtkSpinButton,
because otherwise when you're in the middle of
scrolling the window you can involuntarily
scroll over the widgets and change its values.

This problem can be seen in applications like
pavucontrol, gnome-control-center audio page,
gnome-tweaks, devhelp prefs, and so on.

However this patch allows to scroll the child
widgets when the GtkScrolledWindow itself cannot
be scrolled (i.e. when no scrollbars are shown
because window height exceeds the content).

Fixes issue #3092