listbase: Make sure items are in rubberband before adding
The current implementation of get_items_in_rect uses the
item's coordinates to decides whether an item is in the
rubberband's rectangle or not.  This is usually fine,
except if the widget is not the same size as the list item
(i.e. if there are margins around the widget).  The widget is
visible to the user (not the item), so let's make sure that
the widget is really inside of the rubberband.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2436