Do not use new G_APPLICATION_DEFAULT_FLAGS enum value
The merge request !5235 switched tests and examples from
G_APPLICATION_FLAGS_NONE, which has been deprecated in GLib 2.74, to
G_APPLICATION_DEFAULT_FLAGS. Sadly, it was done unconditionally, which
means we'd have to bump the required version of GLib.

To avoid that, let's just use the numeric value of the enum member.